+++ src/app/release.js
@@ -18,19 +18,61 @@ export class ReleaseItem {
 
 export class Release {
   /**
-   * @param {String} artist
-   * @param {String} title
-   * @param {String} label
-   * @param {Date} date
-   * @param {Array} tracks
-   * @param {String} url
-   * @param {String} image
-   * @param {Array} keywords
+   * @type {ReleaseItem}
    */
-  constructor(artist, title, label, date, tracks, url, image, keywords) {
+  releaseItem;
+
+  /**
+   * @type {string}
+   */
+  label;
+
+  /**
+   * @type {Date}
+   */
+  published;
+
+  /**
+   * @type {Date}
+   */
+  modified;
+
+  /**
+   * @type {Array.<Track>}
+   */
+  tracks;
+
+  /**
+   * @type {Number}
+   */
+  tracksQty;
+
+  /**
+   * @type {string}
+   */
+  image;
+
+  /**
+   * @type {Array.<string>}
+   */
+  keywords;
+
+  /**
+   * @param {string} artist
+   * @param {string} title
+   * @param {string} label
+   * @param {Date} datePublished
+   * @param {Date} dateModified
+   * @param {Array.<Track>} tracks
+   * @param {string} url
+   * @param {string} image
+   * @param {Array.<string>} keywords
+   */
+  constructor(artist, title, label, datePublished, dateModified, tracks, url, image, keywords) {
     this.releaseItem = new ReleaseItem(url, artist, title);
     this.label = label;
-    this.date = date;
+    this.published = datePublished;
+    this.modified = dateModified;
     this.tracks = tracks;
     this.tracksQty = tracks.length;
     this.image = image;
@@ -49,75 +91,15 @@ export class Release {
     return this.releaseItem.title;
   }
 
-  /**
-   * @param {Object} TralbumData
-   * @param {Object} BandData
-   * @param {Object} SchemaData
-   * @param {Object} coverSrc
-   * @returns {Release}
-   */
-  static fromBandcampData(TralbumData, BandData, SchemaData, coverSrc) {
-    const { artist, current, url } = TralbumData;
-    const { title, publish_date } = current;
-    const { keywords } = SchemaData;
-    const tracks = TralbumData.trackinfo.map(track => new Track(
-      track.track_num,
-      track.title,
-      durationFromSeconds(Math.trunc(track.duration))
-    ));
-    const labelName = BandData.name;
-    const label = artist === labelName ? `Not On Label (${labelName} Self-released)` : labelName;
-
-    return new Release(
-      artist,
-      title,
-      label,
-      new Date(publish_date),
-      tracks,
-      url,
-      coverSrc.big,
-      keywords
-    );
-  }
-
-  /**
-   * @param {Object} schema
-   * @returns {Release}
-   */
-  static fromBandcampSchema(schema) {
-    const artist = schema.byArtist.name;
-    const title = schema.name;
-    const label = schema.publisher.name;
-    const date = new Date(schema.datePublished);
-    const tracks = schema.track.itemListElement.map(track => new Track(
-      track.position,
-      track.item.name,
-      TrackDuration.fromDuration(track.item.duration)
-    ));
-    const url = schema.mainEntityOfPage;
-    const image = schema.image;
-    const keywords = schema.keywords;
-
-    return new Release(
-      artist,
-      title,
-      label,
-      date,
-      tracks,
-      url,
-      image,
-      keywords
-    );
-  }
-
-  toObject() {
+  toStorageObject() {
     return {
       uuid: this.releaseItem.uuid,
       artist: this.releaseItem.artist,
       title: this.releaseItem.title,
       url: this.releaseItem.url,
       label: this.label,
-      date: this.date.toISOString(),
+      published: this.published.toISOString(),
+      modified: this.modified.toISOString(),
       tracks: this.tracks,
       image: this.image,
       keywords: this.keywords
@@ -129,18 +111,31 @@ export class Release {
    * @param {Object} obj - A simple object.
    * @returns {Release} An instance of the Release class.
    */
-  static fromObject(obj) {
+  static fromStorageObject(obj) {
     if (!obj.url || !obj.tracks) {
       throw new Error('Cannot create Release object from object', obj);
     }
 
-    const tracks = obj.tracks.map(trackData => Track.fromObject(trackData));
+    const tracks = obj.tracks.map(trackData => new Track(
+      trackData.num,
+      trackData.title,
+      trackData.duration
+    ));
+
+    if (!obj.published && !obj.date) {
+      throw new Error("Missing published or date property");
+    }
+
+    if (!obj.modified && !obj.date) {
+      throw new Error("Missing published or date property");
+    }
 
     return new Release(
       obj.artist,
       obj.title,
       obj.label,
-      new Date(obj.date),
+      new Date(obj.published ?? obj.date),
+      new Date(obj.modified ?? obj.date),
       tracks,
       obj.url,
       obj.image,
@@ -192,19 +187,6 @@ export class Track {
     this.num = num;
     this.title = title;
     this.duration = duration;
-  }
-
-  /**
-   * Create an instance of the Track class from the object.
-   * @param {Object} obj - A simple object.
-   * @returns {Track} An instance of the Track class.
-   */
-  static fromObject(obj) {
-    return new Track(
-      obj.num,
-      obj.title,
-      obj.duration
-    );
   }
 }